refactored compiler tests for collections and dictionaries
